Break apart processor tests some
